Show a portal that is missing a title to force a load.
This helps often, but not always.

IITC keeps a cache of recently loaded portals which includes memory limits (it
predates things like WeakMap).

Normally, which portal is currently in the portal view will stay in the cache,
regardless.  However, if there are enough portals on the map and multiple
moved portals with links, it is possible to get into a state where there will
always be at least one missing title.

In those cases, reducing the size of the map (browser window), and perhaps
careful moving around of the map while loading to avoid problem portals, it
might be possible to get past a problem area.  But, as is, it is still
possible to get stuck in an area.

In some cases, the best option is to go out and destroy links.